+++ src/piddiplatsch/core/processing.py
@@ -11,13 +11,13 @@ class BaseProcessor:
         self,
         handle_backend: HandleAPIProtocol | None = None,
         retries: int = 0,
-        dry_run: bool = False,
+        publish: bool = False,
         project: str | None = None,
         handle_profile: str | None = None,
         handle_output_filename: str | None = None,
     ):
         self.handle_backend: HandleAPIProtocol = handle_backend or HandleAPI(
-            dry_run=dry_run,
+            publish=publish,
             project=project,
             handle_profile=handle_profile,
             output_filename=handle_output_filename,